pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "neptune-osint"
version = "1.0.0"
authors = ["krispeckt"]
description = "🪼 A best OSINT application written in python"

[tool.poetry.dependencies]
pip = "^24.3"
python = "^3.12"
anytree = "^2.12.1"
python-whois = { git = "https://github.com/richardpenman/whois.git" }
bs4 = "^0.0.2"
dnspython = "^2.7.0"
fake-useragent = "^2.0.3"
